ref(all): Move rollup configs into scripts folder, update pkg paths
