+++ package.json
@@ -24,8 +24,8 @@
     "lint": "./node_modules/.bin/eslint --ext \".ts, .tsx\" src/",
     "lint:fix": "./node_modules/.bin/eslint --ext \".ts, .tsx\" src/ --fix",
     "build": "yarn lint && yarn build:bundles && yarn build:tabs && yarn build:docs",
-    "build:bundles": "./node_modules/.bin/tsc && rollup --config rollup.bundles.config.js",
-    "build:tabs": "./node_modules/.bin/tsc && rollup --config rollup.tabs.config.js",
+    "build:bundles": "./node_modules/.bin/tsc && rollup --config scripts/rollup/configs/bundles.js",
+    "build:tabs": "./node_modules/.bin/tsc && rollup --config scripts/rollup/configs/tabs.js",
     "build:docs": "node ./scripts/docs/index.js",
     "serve": "http-server --cors=* -c-1 -p 8022 ./build",
     "dev": "yarn build:bundles && yarn build:tabs && yarn serve",
